feat(ux/agents/builder): refonte UX paramètres modèle — sélecteur + params inline + retrait pipeline tiers
- Remplace le bouton « Modèle » (qui basculait vers sous-panneau ModelPanel)
  par 2 dropdowns Provider + Model directement dans la page principale du
  builder (pattern flow linéaire, cohérent avec commit a98a8d6cd panel
  Paramètres conversation).
- Ajoute inline la grille de paramètres essentiels (Créativité, Resend Files,
  Thinking, Recherche web — selon provider) + accordéon « Paramètres avancés »
  pour les autres params (Top P, Max Tokens, etc.). Pattern identique au
  panel Paramètres conversation. Description du toggle Recherche web alignée
  sur les clés natives (com_endpoint_anthropic_use_web_search /
  com_endpoint_openai_use_web_search) pour symétrie cognitive avec le panel
  conversation.
- Retire l'accordéon « Recherche web » qui pointait vers le pipeline tiers
  LibreChat (Serper/Tavily/Firecrawl/Jina) : trop technique pour les users
  Vermeer. Pipeline tiers cible Admin Panel ClickHouse en V2 (cf. décision
  5.4 CLAUDE.md). La capability backend `web_search` reste active dans le
  yaml, le tool reste exécutable si présent dans agent.tools[] (0 agent
  concerné en base actuelle, vérifié via mongosh).
- Applique le gating endpoint custom (endpointType !== EModelEndpoint.custom)
  sur le toggle web_search inline, par symétrie avec a98a8d6cd. Évite un 400
  garanti sur French Models/Featherless. useEffect de nettoyage retire le
  param du form si l'user bascule vers custom (pattern Panel.tsx:177-229).
- Supprime ModelPanel.tsx + Panel.model enum + Search/{Form,Action}.tsx
  (dead code après refonte ; récupérables via git si réutilisation V2).
  ApiKeyDialog.tsx + InputSection.tsx conservés (consommés hors builder
  par Chat/Input/ToolDialogs.tsx).
